fix eslint and formatting
